requirements.txt
Flask
flask-cors
requests
python-telegram-bot
python-dotenv
pytz
matplotlib
pandas
numpy
python-telegram-bot==13.15
